+++ lottie/src/main/java/com/airbnb/lottie/animation/keyframe/TransformKeyframeAnimation.java
@@ -15,10 +15,16 @@ import static com.airbnb.lottie.LottieProperty.TRANSFORM_OPACITY;
 import static com.airbnb.lottie.LottieProperty.TRANSFORM_POSITION;
 import static com.airbnb.lottie.LottieProperty.TRANSFORM_ROTATION;
 import static com.airbnb.lottie.LottieProperty.TRANSFORM_SCALE;
+import static com.airbnb.lottie.LottieProperty.TRANSFORM_SKEW;
+import static com.airbnb.lottie.LottieProperty.TRANSFORM_SKEW_ANGLE;
 import static com.airbnb.lottie.LottieProperty.TRANSFORM_START_OPACITY;
 
 public class TransformKeyframeAnimation {
   private final Matrix matrix = new Matrix();
+  private final Matrix skewMatrix1;
+  private final Matrix skewMatrix2;
+  private final Matrix skewMatrix3;
+  private final float[] skewValues;
 
   private boolean isIdentity;
   private final BaseKeyframeAnimation<PointF, PointF> anchorPoint;
@@ -26,6 +32,8 @@ public class TransformKeyframeAnimation {
   private final BaseKeyframeAnimation<ScaleXY, ScaleXY> scale;
   private final BaseKeyframeAnimation<Float, Float> rotation;
   private final BaseKeyframeAnimation<Integer, Integer> opacity;
+  @Nullable private final FloatKeyframeAnimation skew;
+  @Nullable private final FloatKeyframeAnimation skewAngle;
 
   // Used for repeaters
   @Nullable private final BaseKeyframeAnimation<?, Float> startOpacity;
@@ -38,11 +46,30 @@ public class TransformKeyframeAnimation {
       position = null;
       scale = null;
       rotation = null;
+      skew = null;
+      skewAngle = null;
+      skewMatrix1 = null;
+      skewMatrix2 = null;
+      skewMatrix3 = null;
+      skewValues = null;
     } else  {
       anchorPoint = animatableTransform.getAnchorPoint().createAnimation();
       position = animatableTransform.getPosition().createAnimation();
       scale = animatableTransform.getScale().createAnimation();
       rotation = animatableTransform.getRotation().createAnimation();
+      skew = animatableTransform.getSkew() == null ? null : (FloatKeyframeAnimation) animatableTransform.getSkew().createAnimation();
+      if (skew != null) {
+        skewMatrix1 = new Matrix();
+        skewMatrix2 = new Matrix();
+        skewMatrix3 = new Matrix();
+        skewValues = new float[9];
+      } else {
+        skewMatrix1 = null;
+        skewMatrix2 = null;
+        skewMatrix3 = null;
+        skewValues = null;
+      }
+      skewAngle = animatableTransform.getSkewAngle() == null ? null : (FloatKeyframeAnimation) animatableTransform.getSkewAngle().createAnimation();
     }
     opacity = animatableTransform.getOpacity().createAnimation();
     if (animatableTransform.getStartOpacity() != null) {
@@ -73,6 +100,12 @@ public class TransformKeyframeAnimation {
     layer.addAnimation(position);
     layer.addAnimation(scale);
     layer.addAnimation(rotation);
+    if (skew != null) {
+      layer.addAnimation(skew);
+    }
+    if (skewAngle != null) {
+      layer.addAnimation(skewAngle);
+    }
   }
 
   public void addListener(final BaseKeyframeAnimation.AnimationListener listener) {
@@ -91,6 +124,12 @@ public class TransformKeyframeAnimation {
     position.addUpdateListener(listener);
     scale.addUpdateListener(listener);
     rotation.addUpdateListener(listener);
+    if (skew != null) {
+      skew.addUpdateListener(listener);
+    }
+    if (skewAngle != null) {
+      skewAngle.addUpdateListener(listener);
+    }
   }
 
   public void  setProgress(float progress) {
@@ -109,6 +148,12 @@ public class TransformKeyframeAnimation {
     position.setProgress(progress);
     scale.setProgress(progress);
     rotation.setProgress(progress);
+    if (skew != null) {
+      skew.setProgress(progress);
+    }
+    if (skewAngle != null) {
+      skewAngle.setProgress(progress);
+    }
   }
 
   public BaseKeyframeAnimation<?, Integer> getOpacity() {
@@ -122,7 +167,6 @@ public class TransformKeyframeAnimation {
   @Nullable public BaseKeyframeAnimation<?, Float> getEndOpacity() {
     return endOpacity;
   }
-
 
   public Matrix getMatrix() {
     if (isIdentity) {
@@ -139,6 +183,36 @@ public class TransformKeyframeAnimation {
       matrix.preRotate(rotation);
     }
 
+    if (skew != null) {
+      float mCos = skewAngle == null ? 0f : (float) Math.cos(Math.toRadians(-skewAngle.getFloatValue() + 90));
+      float mSin = skewAngle == null ? 0f : (float) Math.sin(Math.toRadians(-skewAngle.getFloatValue() + 90));
+      float aTan = (float) Math.tan(Math.toRadians(skew.getFloatValue()));
+      clearSkewValues();
+      skewValues[0] = mCos;
+      skewValues[1] = mSin;
+      skewValues[3] = -mSin;
+      skewValues[4] = mCos;
+      skewValues[8] = 1f;
+      skewMatrix1.setValues(skewValues);
+      clearSkewValues();
+      skewValues[0] = 1f;
+      skewValues[3] = aTan;
+      skewValues[4] = 1f;
+      skewValues[8] = 1f;
+      skewMatrix2.setValues(skewValues);
+      clearSkewValues();
+      skewValues[0] = mCos;
+      skewValues[1] = -mSin;
+      skewValues[3] = mSin;
+      skewValues[4] = mCos;
+      skewValues[8] = 1;
+      skewMatrix3.setValues(skewValues);
+      skewMatrix2.preConcat(skewMatrix1);
+      skewMatrix3.preConcat(skewMatrix2);
+
+      matrix.preConcat(skewMatrix3);
+    }
+
     ScaleXY scaleTransform = this.scale.getValue();
     if (scaleTransform.getScaleX() != 1f || scaleTransform.getScaleY() != 1f) {
       matrix.preScale(scaleTransform.getScaleX(), scaleTransform.getScaleY());
@@ -148,7 +222,14 @@ public class TransformKeyframeAnimation {
     if (anchorPoint.x != 0 || anchorPoint.y != 0) {
       matrix.preTranslate(-anchorPoint.x, -anchorPoint.y);
     }
+
     return matrix;
+  }
+
+  private void clearSkewValues() {
+    for (int i = 0; i < 9; i++) {
+      skewValues[i] = 0f;
+    }
   }
 
   /**
@@ -193,6 +274,10 @@ public class TransformKeyframeAnimation {
       startOpacity.setValueCallback((LottieValueCallback<Float>) callback);
     } else if (property == TRANSFORM_END_OPACITY && endOpacity != null) {
       endOpacity.setValueCallback((LottieValueCallback<Float>) callback);
+    } else if (property == TRANSFORM_SKEW && skew != null) {
+      skew.setValueCallback((LottieValueCallback<Float>) callback);
+    } else if (property == TRANSFORM_SKEW_ANGLE && skewAngle != null) {
+      skewAngle.setValueCallback((LottieValueCallback<Float>) callback);
     } else {
       return false;
     }